+++ generator/example/many_to_many.jorm.dart
@@ -15,6 +15,12 @@ abstract class _TodoListBean implements Bean<TodoList> {
 
   final StrField description = new StrField('description');
 
+  Map<String, Field> _fields;
+  Map<String, Field> get fields => _fields ??= {
+        id.name: id,
+        id1.name: id1,
+        description.name: description,
+      };
   TodoList fromMap(Map map) {
     TodoList model = new TodoList();
 
@@ -25,12 +31,20 @@ abstract class _TodoListBean implements Bean<TodoList> {
     return model;
   }
 
-  List<SetColumn> toSetColumns(TodoList model, [bool update = false]) {
+  List<SetColumn> toSetColumns(TodoList model,
+      {bool update = false, Set<String> only}) {
     List<SetColumn> ret = [];
 
-    ret.add(id.set(model.id));
-    ret.add(id1.set(model.id1));
-    ret.add(description.set(model.description));
+    if (only == null) {
+      ret.add(id.set(model.id));
+      ret.add(id1.set(model.id1));
+      ret.add(description.set(model.description));
+    } else {
+      if (only.contains(id.name)) ret.add(id.set(model.id));
+      if (only.contains(id1.name)) ret.add(id1.set(model.id1));
+      if (only.contains(description.name))
+        ret.add(description.set(model.description));
+    }
 
     return ret;
   }
@@ -60,9 +74,10 @@ abstract class _TodoListBean implements Bean<TodoList> {
   }
 
   Future<int> update(TodoList model,
-      {bool cascade: false, bool associate: false}) async {
-    final Update update =
-        updater.where(this.id.eq(model.id)).setMany(toSetColumns(model));
+      {bool cascade: false, bool associate: false, Set<String> only}) async {
+    final Update update = updater
+        .where(this.id.eq(model.id))
+        .setMany(toSetColumns(model, only: only));
     final ret = execUpdate(update);
     if (cascade) {
       TodoList newModel;
@@ -85,11 +100,6 @@ abstract class _TodoListBean implements Bean<TodoList> {
     return model;
   }
 
-  Future<List<TodoList>> findWhere(Expression exp) async {
-    final Find find = finder.where(exp);
-    return await (await execFind(find)).toList();
-  }
-
   Future<int> remove(String id, [bool cascade = false]) async {
     if (cascade) {
       final TodoList newModel = await find(id);
@@ -105,10 +115,6 @@ abstract class _TodoListBean implements Bean<TodoList> {
       remove.or(this.id.eq(model.id));
     }
     return execRemove(remove);
-  }
-
-  Future<int> removeWhere(Expression exp) async {
-    return execRemove(remover.where(exp));
   }
 
   Future preload(TodoList model, {bool cascade: false}) async {
@@ -128,6 +134,11 @@ abstract class _CategoryBean implements Bean<Category> {
 
   final StrField name = new StrField('name');
 
+  Map<String, Field> _fields;
+  Map<String, Field> get fields => _fields ??= {
+        id.name: id,
+        name.name: name,
+      };
   Category fromMap(Map map) {
     Category model = new Category();
 
@@ -137,11 +148,17 @@ abstract class _CategoryBean implements Bean<Category> {
     return model;
   }
 
-  List<SetColumn> toSetColumns(Category model, [bool update = false]) {
+  List<SetColumn> toSetColumns(Category model,
+      {bool update = false, Set<String> only}) {
     List<SetColumn> ret = [];
 
-    ret.add(id.set(model.id));
-    ret.add(name.set(model.name));
+    if (only == null) {
+      ret.add(id.set(model.id));
+      ret.add(name.set(model.name));
+    } else {
+      if (only.contains(id.name)) ret.add(id.set(model.id));
+      if (only.contains(name.name)) ret.add(name.set(model.name));
+    }
 
     return ret;
   }
@@ -170,9 +187,10 @@ abstract class _CategoryBean implements Bean<Category> {
   }
 
   Future<int> update(Category model,
-      {bool cascade: false, bool associate: false}) async {
-    final Update update =
-        updater.where(this.id.eq(model.id)).setMany(toSetColumns(model));
+      {bool cascade: false, bool associate: false, Set<String> only}) async {
+    final Update update = updater
+        .where(this.id.eq(model.id))
+        .setMany(toSetColumns(model, only: only));
     final ret = execUpdate(update);
     if (cascade) {
       Category newModel;
@@ -195,11 +213,6 @@ abstract class _CategoryBean implements Bean<Category> {
     return model;
   }
 
-  Future<List<Category>> findWhere(Expression exp) async {
-    final Find find = finder.where(exp);
-    return await (await execFind(find)).toList();
-  }
-
   Future<int> remove(String id, [bool cascade = false]) async {
     if (cascade) {
       final Category newModel = await find(id);
@@ -215,10 +228,6 @@ abstract class _CategoryBean implements Bean<Category> {
       remove.or(this.id.eq(model.id));
     }
     return execRemove(remove);
-  }
-
-  Future<int> removeWhere(Expression exp) async {
-    return execRemove(remover.where(exp));
   }
 
   Future preload(Category model, {bool cascade: false}) async {
@@ -240,6 +249,12 @@ abstract class _PivotBean implements Bean<Pivot> {
 
   final StrField categoryId = new StrField('category_id');
 
+  Map<String, Field> _fields;
+  Map<String, Field> get fields => _fields ??= {
+        todolistId.name: todolistId,
+        todolistId1.name: todolistId1,
+        categoryId.name: categoryId,
+      };
   Pivot fromMap(Map map) {
     Pivot model = new Pivot();
 
@@ -250,12 +265,22 @@ abstract class _PivotBean implements Bean<Pivot> {
     return model;
   }
 
-  List<SetColumn> toSetColumns(Pivot model, [bool update = false]) {
+  List<SetColumn> toSetColumns(Pivot model,
+      {bool update = false, Set<String> only}) {
     List<SetColumn> ret = [];
 
-    ret.add(todolistId.set(model.todolistId));
-    ret.add(todolistId1.set(model.todolistId1));
-    ret.add(categoryId.set(model.categoryId));
+    if (only == null) {
+      ret.add(todolistId.set(model.todolistId));
+      ret.add(todolistId1.set(model.todolistId1));
+      ret.add(categoryId.set(model.categoryId));
+    } else {
+      if (only.contains(todolistId.name))
+        ret.add(todolistId.set(model.todolistId));
+      if (only.contains(todolistId1.name))
+        ret.add(todolistId1.set(model.todolistId1));
+      if (only.contains(categoryId.name))
+        ret.add(categoryId.set(model.categoryId));
+    }
 
     return ret;
   }
@@ -274,15 +299,6 @@ abstract class _PivotBean implements Bean<Pivot> {
   Future<dynamic> insert(Pivot model) async {
     final Insert insert = inserter.setMany(toSetColumns(model));
     return execInsert(insert);
-  }
-
-  Future<List<Pivot>> findWhere(Expression exp) async {
-    final Find find = finder.where(exp);
-    return await (await execFind(find)).toList();
-  }
-
-  Future<int> removeWhere(Expression exp) async {
-    return execRemove(remover.where(exp));
   }
 
   Future<List<Pivot>> findByTodoList(String todolistId, int todolistId1,