+++ orm/example/associations/one_to_one/one_to_one.jorm.dart
@@ -13,6 +13,11 @@ abstract class _UserBean implements Bean<User> {
 
   final StrField name = new StrField('name');
 
+  Map<String, Field> _fields;
+  Map<String, Field> get fields => _fields ??= {
+        id.name: id,
+        name.name: name,
+      };
   User fromMap(Map map) {
     User model = new User();
 
@@ -22,11 +27,17 @@ abstract class _UserBean implements Bean<User> {
     return model;
   }
 
-  List<SetColumn> toSetColumns(User model, [bool update = false]) {
+  List<SetColumn> toSetColumns(User model,
+      {bool update = false, Set<String> only}) {
     List<SetColumn> ret = [];
 
-    ret.add(id.set(model.id));
-    ret.add(name.set(model.name));
+    if (only == null) {
+      ret.add(id.set(model.id));
+      ret.add(name.set(model.name));
+    } else {
+      if (only.contains(id.name)) ret.add(id.set(model.id));
+      if (only.contains(name.name)) ret.add(name.set(model.name));
+    }
 
     return ret;
   }
@@ -53,9 +64,10 @@ abstract class _UserBean implements Bean<User> {
   }
 
   Future<int> update(User model,
-      {bool cascade: false, bool associate: false}) async {
-    final Update update =
-        updater.where(this.id.eq(model.id)).setMany(toSetColumns(model));
+      {bool cascade: false, bool associate: false, Set<String> only}) async {
+    final Update update = updater
+        .where(this.id.eq(model.id))
+        .setMany(toSetColumns(model, only: only));
     final ret = execUpdate(update);
     if (cascade) {
       User newModel;
@@ -80,11 +92,6 @@ abstract class _UserBean implements Bean<User> {
     return model;
   }
 
-  Future<List<User>> findWhere(Expression exp) async {
-    final Find find = finder.where(exp);
-    return await (await execFind(find)).toList();
-  }
-
   Future<int> remove(String id, [bool cascade = false]) async {
     if (cascade) {
       final User newModel = await find(id);
@@ -100,10 +107,6 @@ abstract class _UserBean implements Bean<User> {
       remove.or(this.id.eq(model.id));
     }
     return execRemove(remove);
-  }
-
-  Future<int> removeWhere(Expression exp) async {
-    return execRemove(remover.where(exp));
   }
 
   Future preload(User model, {bool cascade: false}) async {
@@ -133,6 +136,12 @@ abstract class _AddressBean implements Bean<Address> {
 
   final StrField userId = new StrField('user_id');
 
+  Map<String, Field> _fields;
+  Map<String, Field> get fields => _fields ??= {
+        id.name: id,
+        street.name: street,
+        userId.name: userId,
+      };
   Address fromMap(Map map) {
     Address model = new Address();
 
@@ -143,12 +152,19 @@ abstract class _AddressBean implements Bean<Address> {
     return model;
   }
 
-  List<SetColumn> toSetColumns(Address model, [bool update = false]) {
+  List<SetColumn> toSetColumns(Address model,
+      {bool update = false, Set<String> only}) {
     List<SetColumn> ret = [];
 
-    ret.add(id.set(model.id));
-    ret.add(street.set(model.street));
-    ret.add(userId.set(model.userId));
+    if (only == null) {
+      ret.add(id.set(model.id));
+      ret.add(street.set(model.street));
+      ret.add(userId.set(model.userId));
+    } else {
+      if (only.contains(id.name)) ret.add(id.set(model.id));
+      if (only.contains(street.name)) ret.add(street.set(model.street));
+      if (only.contains(userId.name)) ret.add(userId.set(model.userId));
+    }
 
     return ret;
   }
@@ -166,9 +182,10 @@ abstract class _AddressBean implements Bean<Address> {
     return execInsert(insert);
   }
 
-  Future<int> update(Address model) async {
-    final Update update =
-        updater.where(this.id.eq(model.id)).setMany(toSetColumns(model));
+  Future<int> update(Address model, {Set<String> only}) async {
+    final Update update = updater
+        .where(this.id.eq(model.id))
+        .setMany(toSetColumns(model, only: only));
     return execUpdate(update);
   }
 
@@ -176,11 +193,6 @@ abstract class _AddressBean implements Bean<Address> {
       {bool preload: false, bool cascade: false}) async {
     final Find find = finder.where(this.id.eq(id));
     return await execFindOne(find);
-  }
-
-  Future<List<Address>> findWhere(Expression exp) async {
-    final Find find = finder.where(exp);
-    return await (await execFind(find)).toList();
   }
 
   Future<int> remove(String id) async {
@@ -194,10 +206,6 @@ abstract class _AddressBean implements Bean<Address> {
       remove.or(this.id.eq(model.id));
     }
     return execRemove(remove);
-  }
-
-  Future<int> removeWhere(Expression exp) async {
-    return execRemove(remover.where(exp));
   }
 
   Future<Address> findByUser(String userId,