+++ query/example/create_table/one_to_many/example.dart
@@ -185,7 +185,6 @@ main() async {
 
   {
     Post post = await getPostByIdRelated(post1Id);
-
     print(post);
   }