+++ orm/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.2
+
++ Upgraded to `jaguar_query` `2.1.5`
+
 ## 2.1.1
 
 + Dart2 support