+++ query/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.5
+
++ Added `fields` to `Bean`
+
+## 2.1.4
+
++ Added `findOneWhere`, `findWhere` and `removeWhere` to `Bean`
+
 ## 2.1.2
 
 + Added `parseValue` to adapter